Use 32-bit integers due to xtgeo
Usage of uint16 is a valid optimization for the region parameter, but
xtgeo is using numpy with a fill value that overflows in uint16.

As long as xtgeo does not support uint16 explicitly, we need to
resort to 32-bits.

https://github.com/equinor/xtgeo/blob/69becba852ace157c8610f9b2edcfb582e7d96f9/src/xtgeo/common/constants.py#L13

This triggers a warning from numpy stating that this condition will lead
to an error in the future

+++ tests/jobs/localisation/test_integration.py
@@ -358,10 +358,10 @@ def create_region_parameter(filename, grid):
     region_param = xtgeo.GridProperty(
         grid, name=region_param_name, discrete=True, values=1
     )
-    region_param.dtype = np.uint16
+    region_param.dtype = np.int32
     region_param.codes = region_code_names
     (nx, ny, nz) = grid.dimensions
-    values = np.zeros((nx, ny, nz), dtype=np.uint16)
+    values = np.zeros((nx, ny, nz), dtype=np.int32)
     values[:, :, :] = 0
     for k, j, i in itertools.product(range(nz), range(ny), range(nx)):
         if 0 <= i <= nx / 2 and 0 <= j <= ny / 2:
